Moved navigation out to separate menu item
