+++ app/components/gh-nav-menu/settings.hbs
@@ -9,45 +9,45 @@
 <div class="gh-nav-top">
     <ul class="gh-nav-list gh-nav-settings-main">
         <li class="relative">
-            <LinkTo @route="dashboard" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "arrow-left"}}Back to Dashboard</LinkTo>
+            <LinkTo @route="dashboard" @alt="Back to Dashboard" @title="Back to Dashboard" data-test-nav="dashboard">{{svg-jar "arrow-left"}}Back to Dashboard</LinkTo>
         </li>
     </ul>
     <ul class="gh-nav-list">
         <li class="gh-nav-list-h">Website</li>
         <li class="relative">
-            <LinkTo @route="settings.general" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "page"}}General</LinkTo>
+            <LinkTo @route="settings.general" @alt="General" @title="General" data-test-nav="settings">{{svg-jar "page"}}General</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.theme" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "paintbrush"}}Theme</LinkTo>
+            <LinkTo @route="settings.theme" @alt="Theme" @title="Theme" data-test-nav="theme">{{svg-jar "paintbrush"}}Theme</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.design" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "paintbrush"}}Branding</LinkTo>
+            <LinkTo @route="settings.design" @alt="Branding" @title="Branding" data-test-nav="design">{{svg-jar "paintbrush"}}Branding</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.design" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "house"}}Navigation</LinkTo>
+            <LinkTo @route="settings.navigation" @alt="Navigation" @title="Navigation" data-test-nav="navigation">{{svg-jar "house"}}Navigation</LinkTo>
         </li>
     </ul>
 
     <ul class="gh-nav-list">
         <li class="gh-nav-list-h">Members</li>
         <li class="relative">
-            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Portal</LinkTo>
+            <LinkTo @route="settings.members" @alt="Portal" @title="Portal" data-test-nav="members-settings">{{svg-jar "members"}}Portal</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Email newsletter</LinkTo>
+            <LinkTo @route="settings.members" @alt="Email" @title="Email" data-test-nav="members-settings">{{svg-jar "members"}}Email newsletter</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Payments</LinkTo>
+            <LinkTo @route="settings.members" @alt="Payments" @title="Payments" data-test-nav="members-settings">{{svg-jar "members"}}Payments</LinkTo>
         </li>
     </ul>
 
     <ul class="gh-nav-list">
         <li class="gh-nav-list-h">Extensions</li>
         <li class="relative">
-            <LinkTo @route="settings.code-injection" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "brackets"}}Code injection</LinkTo>
+            <LinkTo @route="settings.code-injection" @alt="Code injection" @title="Code injection" data-test-nav="code-injection">{{svg-jar "brackets"}}Code injection</LinkTo>
         </li>
         <li class="relative">
-            <LinkTo @route="settings.labs" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "labs"}}Labs</LinkTo>
+            <LinkTo @route="settings.labs" @alt="Labs" @title="Labs" data-test-nav="labs">{{svg-jar "labs"}}Labs</LinkTo>
         </li>
     </ul>
 </div>
