+++ app/templates/settings/design-loading.hbs
@@ -1,9 +1,7 @@
 <section class="gh-canvas">
     <header class="gh-canvas-header">
         <h2 class="gh-canvas-title" data-test-screen-title>
-            <LinkTo @route="settings">Settings</LinkTo>
-            <span>{{svg-jar "arrow-right"}}</span>
-            Design
+            Branding
         </h2>
         <section class="view-actions">
             <GhTaskButton @task={{save}} @class="gh-btn gh-btn-blue gh-btn-icon" @disabled={{true}} />
