+++ app/templates/settings/design.hbs
@@ -1,9 +1,7 @@
 <section class="gh-canvas">
     <GhCanvasHeader class="gh-canvas-header">
         <h2 class="gh-canvas-title" data-test-screen-title>
-            <LinkTo @route="settings">Settings</LinkTo>
-            <span>{{svg-jar "arrow-right"}}</span>
-            Design
+            Branding
         </h2>
         <section class="view-actions">
             <GhTaskButton @task={{this.save}} @class="gh-btn gh-btn-blue gh-btn-icon" data-test-save-button={{true}} />
@@ -161,60 +159,6 @@
                     </GhFormGroup>
                 </div>
             </div>
-        </div>
-
-        <div class="gh-setting-header">Navigation</div>
-        <div class="gh-blognav-container pa5 pt6 bg-grouped-table shadow-1 br3">
-            <form id="settings-navigation" class="gh-blognav" novalidate="novalidate">
-                <SortableObjects @sortableObjectList={{this.settings.navigation}} @useSwap={{false}}>
-                    {{#each this.settings.navigation as |navItem index|}}
-                        <DraggableObject @content={{navItem}} @dragHandle=".gh-blognav-grab" @isSortable={{true}}>
-                            <GhNavitem
-                                @navItem={{navItem}}
-                                @baseUrl={{this.blogUrl}}
-                                @addItem={{action "addNavItem"}}
-                                @deleteItem={{action "deleteNavItem"}}
-                                @updateUrl={{action "updateUrl"}}
-                                @updateLabel={{action "updateLabel"}}
-                                data-test-navitem={{index}} />
-                        </DraggableObject>
-                    {{/each}}
-                </SortableObjects>
-                <GhNavitem
-                    @navItem={{this.newNavItem}}
-                    @baseUrl={{this.blogUrl}}
-                    @addItem={{action "addNavItem"}}
-                    @updateUrl={{action "updateUrl"}}
-                    @updateLabel={{action "updateLabel"}}
-                    data-test-navitem="new" />
-            </form>
-        </div>
-
-        <div class="gh-setting-header">Secondary Navigation</div>
-        <div class="gh-blognav-container pa5 pt6 bg-grouped-table shadow-1 br3">
-            <form id="secondary-navigation" class="gh-blognav" novalidate="novalidate">
-                <SortableObjects @sortableObjectList={{this.settings.secondaryNavigation}} @useSwap={{false}}>
-                    {{#each this.settings.secondaryNavigation as |navItem index|}}
-                        <DraggableObject @content={{navItem}} @dragHandle=".gh-blognav-grab" @isSortable={{true}}>
-                            <GhNavitem
-                                @navItem={{navItem}}
-                                @baseUrl={{this.blogUrl}}
-                                @addItem={{action "addNavItem"}}
-                                @deleteItem={{action "deleteNavItem"}}
-                                @updateUrl={{action "updateUrl"}}
-                                @updateLabel={{action "updateLabel"}}
-                                data-test-navitem={{index}} />
-                        </DraggableObject>
-                    {{/each}}
-                </SortableObjects>
-                <GhNavitem
-                    @navItem={{this.newSecondaryNavItem}}
-                    @baseUrl={{this.blogUrl}}
-                    @addItem={{action "addNavItem"}}
-                    @updateUrl={{action "updateUrl"}}
-                    @updateLabel={{action "updateLabel"}}
-                    data-test-navitem="new" />
-            </form>
         </div>
     </section>
 </section>
