--- app/templates/settings/design/uploadtheme.hbs
@@ -1,7 +1,0 @@
-<GhFullscreenModal @modal="upload-theme"
-    @model={{hash
-        themes=themes
-        activate=(route-action 'activateTheme')
-    }}
-    @close={{route-action "cancel"}}
-    @modifier="action wide" />
